+++ app/src/main/java/io/homeassistant/companion/android/vehicle/MainVehicleScreen.kt
@@ -12,9 +12,7 @@ import androidx.car.app.model.Action
 import androidx.car.app.model.GridTemplate
 import androidx.car.app.model.ItemList
 import androidx.car.app.model.Template
-import androidx.lifecycle.DefaultLifecycleObserver
 import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import io.homeassistant.companion.android.BuildConfig
@@ -74,16 +72,8 @@ class MainVehicleScreen(
     private var favoritesList = emptyList<String>()
     private var isLoggedIn: Boolean? = null
     private val domains = mutableSetOf<String>()
-    private var car: Car? = null
-    private var carRestrictionManager: CarUxRestrictionsManager? = null
-    val iDrivingOptimized
-        get() = car?.let {
-            (
-                it.getCarManager(Car.CAR_UX_RESTRICTION_SERVICE) as CarUxRestrictionsManager
-                ).getCurrentCarUxRestrictions().isRequiresDistractionOptimization()
-        } ?: false
 
-    val isAutomotive get() = carContext.packageManager.hasSystemFeature(PackageManager.FEATURE_AUTOMOTIVE)
+    private val isAutomotive get() = carContext.packageManager.hasSystemFeature(PackageManager.FEATURE_AUTOMOTIVE)
 
     init {
         lifecycleScope.launch {
@@ -116,19 +106,6 @@ class MainVehicleScreen(
                 favoriteEntities = allEntities.map { getFavoritesList(it) }
             }
         }
-
-        lifecycle.addObserver(object : DefaultLifecycleObserver {
-
-            override fun onResume(owner: LifecycleOwner) {
-                registerAutomotiveRestrictionListener()
-            }
-
-            override fun onPause(owner: LifecycleOwner) {
-                carRestrictionManager?.unregisterListener()
-                car?.disconnect()
-                car = null
-            }
-        })
     }
 
     override fun onGetTemplate(): Template {
@@ -183,7 +160,7 @@ class MainVehicleScreen(
         return GridTemplate.Builder().apply {
             setTitle(carContext.getString(commonR.string.app_name))
             setHeaderAction(Action.APP_ICON)
-            if (isAutomotive && !iDrivingOptimized && BuildConfig.FLAVOR != "full") {
+            if (isAutomotive && !HaCarAppService().isDrivingOptimized && BuildConfig.FLAVOR != "full") {
                 setActionStrip(nativeModeActionStrip(carContext))
             }
             if (domains.isEmpty()) {
@@ -195,17 +172,17 @@ class MainVehicleScreen(
         }.build()
     }
 
-    private fun registerAutomotiveRestrictionListener() {
+    fun registerAutomotiveRestrictionListener() {
         if (carContext.packageManager.hasSystemFeature(PackageManager.FEATURE_AUTOMOTIVE)) {
             Log.i(TAG, "Register for Automotive Restrictions")
-            car = Car.createCar(carContext)
-            carRestrictionManager =
-                car?.getCarManager(Car.CAR_UX_RESTRICTION_SERVICE) as CarUxRestrictionsManager
+            HaCarAppService().car = Car.createCar(carContext)
+            HaCarAppService().carRestrictionManager =
+                HaCarAppService().car?.getCarManager(Car.CAR_UX_RESTRICTION_SERVICE) as CarUxRestrictionsManager
             val listener =
                 CarUxRestrictionsManager.OnUxRestrictionsChangedListener { restrictions ->
                     invalidate()
                 }
-            carRestrictionManager?.registerListener(listener)
+            HaCarAppService().carRestrictionManager?.registerListener(listener)
         }
     }
